+++ packages/klarna-integration/src/klarnav2/klarnav2.mock.ts
@@ -159,3 +159,32 @@ export function getKlarna(): PaymentMethod {
         clientToken: 'foo',
     };
 }
+
+export function getKlarnaV2UpdateSessionParamsWithOrganizationName(): KlarnaUpdateSessionParams {
+    return {
+        billing_address: {
+            street_address: '12345 Testing Way',
+            city: 'Some City',
+            country: 'DE',
+            given_name: 'Test',
+            family_name: 'Tester',
+            postal_code: '95555',
+            region: 'Berlin',
+            email: 'test@bigcommerce.com',
+            phone: '555-555-5555',
+            organization_name: 'Bigcommerce',
+        },
+        shipping_address: {
+            street_address: '12345 Testing Way',
+            city: 'Some City',
+            country: 'US',
+            given_name: 'Test',
+            family_name: 'Tester',
+            postal_code: '95555',
+            region: 'California',
+            email: 'test@bigcommerce.com',
+            phone: '555-555-5555',
+            organization_name: 'Bigcommerce',
+        },
+    };
+}